various warnings fixed
Ignore-this: 97f57970d2b129af9a802f16c8bee62f

darcs-hash:20090623161308-68901-20b3b9d435f2e573b3297c72382d062847743606.gz

+++ kernel/tr/pstr/pstr_long.h
@@ -202,7 +202,7 @@ public:
 		//no error checks here, TODO - add some assertions
 		//we don't care if we go before the start of string too far
 		--m_pos;
-		if (((int)(--ofs)) < PSTR_LONG_BLK_HDR_SIZE)
+		if (((int)(--ofs)) < (int) PSTR_LONG_BLK_HDR_SIZE)
 		{
 			CHECKP(blk);
 			if (PSTR_LONG_BLK_HDR(blk)->prev_blk != XNULL)
